requirements.txt
netifaces
bluepy
paho-mqtt
delayed